[alerting] Rename alerts to alerting.
