+++ probes/alerting/proto/config_proto_gen.cue
