+++ probes/options/options.go
@@ -25,7 +25,7 @@ import (
 	"github.com/cloudprober/cloudprober/common/iputils"
 	"github.com/cloudprober/cloudprober/logger"
 	"github.com/cloudprober/cloudprober/metrics"
-	"github.com/cloudprober/cloudprober/probes/alerts"
+	"github.com/cloudprober/cloudprober/probes/alerting"
 	configpb "github.com/cloudprober/cloudprober/probes/proto"
 	"github.com/cloudprober/cloudprober/targets"
 	"github.com/cloudprober/cloudprober/targets/endpoint"
@@ -49,7 +49,7 @@ type Options struct {
 	LogMetrics          func(*metrics.EventMetrics)
 	AdditionalLabels    []*AdditionalLabel
 	NegativeTest        bool
-	AlertHandlers       []*alerts.AlertHandler
+	AlertHandlers       []*alerting.AlertHandler
 }
 
 const defaultStatsExtportIntv = 10 * time.Second
@@ -215,7 +215,7 @@ func BuildProbeOptions(p *configpb.ProbeDef, ldLister endpoint.Lister, globalTar
 		if alertConf.GetName() == "" {
 			alertConf.Name = p.GetName()
 		}
-		ah, err := alerts.NewAlertHandler(alertConf, opts.Logger)
+		ah, err := alerting.NewAlertHandler(alertConf, opts.Logger)
 		if err != nil {
 			return nil, fmt.Errorf("error in initializing alert handler: %v", err)
 		}
